+++ hahomematic/client/__init__.py
@@ -15,7 +15,6 @@ from hahomematic.config import CALLBACK_WARN_INTERVAL, RECONNECT_WAIT, WAIT_FOR_
 from hahomematic.const import (
     DATETIME_FORMAT_MILLIS,
     DEFAULT_CUSTOM_ID,
-    DEFAULT_PARAMSETS,
     ENTITY_KEY,
     EVENT_AVAILABLE,
     EVENT_SECONDS_SINCE_LAST_EVENT,
@@ -46,6 +45,7 @@ from hahomematic.support import (
     build_headers,
     build_xml_rpc_uri,
     get_device_address,
+    is_channel_address,
     reduce_args,
     supports_rx_mode,
 )
@@ -554,7 +554,7 @@ class Client(ABC):
     async def put_paramset(
         self,
         channel_address: str,
-        paramset_key: ParamsetKey,
+        paramset_key: ParamsetKey | str,
         values: dict[str, Any],
         wait_for_callback: int | None = WAIT_FOR_CALLBACK,
         rx_mode: CommandRxMode | None = None,
@@ -566,14 +566,22 @@ class Client(ABC):
         Address is usually the channel_address,
         but for bidcos devices there is a master paramset at the device.
         """
+        is_link_call: bool = False
         try:
-            checked_values = (
-                self._check_put_paramset(
-                    channel_address=channel_address, paramset_key=paramset_key, values=values
+            if check_against_pd:
+                check_paramset_key = (
+                    ParamsetKey.LINK
+                    if (
+                        is_link_call := is_channel_address(address=paramset_key)
+                        and isinstance(paramset_key, str)
+                    )
+                    else ParamsetKey(paramset_key)
                 )
-                if check_against_pd
-                else values
-            )
+                checked_values = self._check_put_paramset(
+                    channel_address=channel_address, paramset_key=check_paramset_key, values=values
+                )
+            else:
+                checked_values = values
             _LOGGER.debug(
                 "PUT_PARAMSET: %s, %s, %s", channel_address, paramset_key, checked_values
             )
@@ -587,10 +595,14 @@ class Client(ABC):
                     return set()
             else:
                 await self._proxy.putParamset(channel_address, paramset_key, checked_values)
+
+            if is_link_call:
+                return set()
+
             # store the send value in the last_value_send_cache
             entity_keys = self._last_value_send_cache.add_put_paramset(
                 channel_address=channel_address,
-                paramset_key=paramset_key,
+                paramset_key=ParamsetKey(paramset_key),
                 values=checked_values,
             )
             if wait_for_callback is not None and (
@@ -639,6 +651,7 @@ class Client(ABC):
         value: Any,
         operation: Operations,
     ) -> Any:
+        # Rewrite check for LINK paramset
         """Check a single parameter against paramset descriptions."""
         if parameter_data := self.central.paramset_descriptions.get_parameter_data(
             interface_id=self.interface_id,
@@ -650,7 +663,10 @@ class Client(ABC):
             pd_value_list = (
                 tuple(parameter_data["VALUE_LIST"]) if parameter_data.get("VALUE_LIST") else None
             )
-            if not bool(int(parameter_data["OPERATIONS"]) & operation):
+            if (
+                not bool(pd_operation := int(parameter_data["OPERATIONS"]) & operation)
+                and pd_operation
+            ):
                 raise HaHomematicException(
                     f"Parameter {parameter} does not support the requested operation {operation.value}"
                 )
@@ -698,8 +714,7 @@ class Client(ABC):
         paramsets[address] = {}
         _LOGGER.debug("GET_PARAMSET_DESCRIPTIONS for %s", address)
         for p_key in device_description["PARAMSETS"]:
-            if (paramset_key := ParamsetKey(p_key)) not in DEFAULT_PARAMSETS:
-                continue
+            paramset_key = ParamsetKey(p_key)
             if paramset_description := await self._get_paramset_description(
                 address=address, paramset_key=paramset_key
             ):
@@ -1253,7 +1268,9 @@ async def _track_single_entity_state_change_or_timeout(
 
     channel_address, paramset_key, parameter = entity_key
     if entity := device.get_generic_entity(
-        channel_address=channel_address, parameter=parameter, paramset_key=paramset_key
+        channel_address=channel_address,
+        parameter=parameter,
+        paramset_key=ParamsetKey(paramset_key),
     ):
         if not entity.supports_events:
             _LOGGER.debug(